Removed typescript, rainbowkit, styles
