+++ package.json
@@ -11,15 +11,11 @@
   "dependencies": {
     "@apollo/client": "^3.6.9",
     "@nft/hooks": "^1.0.0-beta.2",
-    "@rainbow-me/rainbowkit": "^0.5.1",
     "@tanstack/react-query": "^4.2.3",
     "ethers": "^5.7.0",
     "next": "^12.1.6",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "wagmi": "^0.6.5"
-  },
-  "devDependencies": {
-    "typescript": "^3.9.10"
   }
 }
